requirements.txt
langchain==0.1.13
langchain-community<0.1,>=0.0.29
langchain_openai==0.1.1
streamlit==1.32.2
tiktoken==0.6.0
fastapi==0.110.0
python-dotenv>=0.21.0
openai==1.14.3
faiss_cpu==1.8.0



